refactor: overlay content components
