+++ src/components/DecentralisedTemplateRenderer/MultiTabs.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent, useContext } from "react";
 import styles from "./multiTabs.scss";
 import { ButtonBorderedBlue } from "../UI/Button";
 import { OverlayContext } from "../../common/contexts/OverlayContext";
-import { OverlayAddressBook } from "./../../components/UI/Overlay";
+import { AddressBook } from "./../../components/UI/Overlay/OverlayContent/AddressBook";
 
 interface MultiTabsProps {
   templates: { id: string; label: string }[];
@@ -19,7 +19,7 @@ export const MultiTabs: FunctionComponent<MultiTabsProps> = ({
 }) => {
   const { setOverlayContent, setOverlayVisible } = useContext(OverlayContext);
   const onOverlayHandler = () => {
-    setOverlayContent(<OverlayAddressBook title="Address Book" data-testid="overlay-addressbook" />);
+    setOverlayContent(<AddressBook title="Address Book" data-testid="overlay-addressbook" />);
     setOverlayVisible(true);
   };
