+++ src/components/UI/Overlay/OverlayContent/Textual.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { OverlayContentBaseStyle } from "../Overlay";
+import { OverlayContent, OverlayContentProps } from "./index";
+import styled from "@emotion/styled";
+
+export const Textual = styled(({ children, ...props }: OverlayContentProps) => {
+  return <OverlayContent {...props}>{children}</OverlayContent>;
+})`
+  ${OverlayContentBaseStyle()}
+`;
